refactor: split admin generate page script
